pyproject.toml
[project]
name = "log-publisher"
version = "0.1.0"
description = "Publish logs to files, or remote servers via REST api, Websocket, or TCP"
authors = [
    {name = "Eliyas-Wubie",email = "115378907+Eliyas-Wubie@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.15"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "pyinstaller (>=6.15.0,<7.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "python-socketio (>=5.13.0,<6.0.0)",
    "websocket-client (>=1.8.0,<2.0.0)",
    "flask (>=3.1.1,<4.0.0)",
    "aiohttp (>=3.12.15,<4.0.0)",
    "websockets (>=15.0.1,<16.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
